+++ packages/deploy-helpers/src/deploy/deploy.ts
@@ -751,10 +751,11 @@ export default async function deploy(
 		accountId,
 		scriptName,
 		workerTag,
-		env: props.env,
 		crons: props.triggers,
 		firstDeploy: !workerExists,
 		routes: props.routes,
+		validated: true,
+		dryRun: false,
 	});
 
 	logger.log("Current Version ID:", versionId);